+++ docs/website/docusaurus.config.js
@@ -31,7 +31,17 @@ const config = {
 
   plugins: [
     path.resolve(__dirname, 'lifecycle/carlin'),
-    ...['appsync-api', 'aws-appsync-nodejs', 'cloud-auth'].map((package) => {
+    ...[
+      'appsync-api',
+      'aws-appsync-nodejs',
+      'cloud-auth',
+      'config',
+      'forms',
+      'i18n-cli',
+      'react-i18n',
+      'test-utils',
+      'ui',
+    ].map((package) => {
       return [
         'docusaurus-plugin-typedoc',
         {
@@ -42,6 +52,7 @@ const config = {
           sidebar: {
             categoryLabel: `@ttoss/${package}`,
           },
+          excludeExternals: true,
         },
       ];
     }),