+++ packages/react-auth/tsup.config.ts
@@ -1,5 +1,3 @@
 import { tsupConfig } from '@ttoss/config';
 
-export const tsup = tsupConfig({
-  noExternal: ['@ttoss/cloud-auth'],
-});
+export const tsup = tsupConfig({});