+++ packages/theme/tsup.config.ts
@@ -1,3 +1,5 @@
 import { tsupConfig } from '@ttoss/config';
 
-export const tsup = tsupConfig();
+export const tsup = tsupConfig({
+  entryPoints: ['src/index.ts', 'src/themes/Bruttal/Bruttal.ts'],
+});