+++ docs/website/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.35.5](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.4...@docs/website@1.35.5) (2023-03-26)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.4](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.3...@docs/website@1.35.4) (2023-03-26)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.3](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.2...@docs/website@1.35.3) (2023-03-22)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.2](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.1...@docs/website@1.35.2) (2023-03-21)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.35.1](https://github.com/ttoss/ttoss/compare/@docs/website@1.35.0...@docs/website@1.35.1) (2023-03-20)
+
+**Note:** Version bump only for package @docs/website
+
+# [1.35.0](https://github.com/ttoss/ttoss/compare/@docs/website@1.34.27...@docs/website@1.35.0) (2023-03-19)
+
+### Features
+
+- improve i18n ([#248](https://github.com/ttoss/ttoss/issues/248)) ([ac8629d](https://github.com/ttoss/ttoss/commit/ac8629d24c83076aa39f8bff5188ee74e1a4de9d))
+
+## [1.34.27](https://github.com/ttoss/ttoss/compare/@docs/website@1.34.26...@docs/website@1.34.27) (2023-03-17)
+
+**Note:** Version bump only for package @docs/website
+
+## [1.34.26](https://github.com/ttoss/ttoss/compare/@docs/website@1.34.25...@docs/website@1.34.26) (2023-03-16)
+
+**Note:** Version bump only for package @docs/website
+
 ## [1.34.25](https://github.com/ttoss/ttoss/compare/@docs/website@1.34.24...@docs/website@1.34.25) (2023-03-15)
 
 **Note:** Version bump only for package @docs/website