+++ packages/google-maps/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.22.10](https://github.com/ttoss/ttoss/compare/@ttoss/google-maps@1.22.9...@ttoss/google-maps@1.22.10) (2023-03-26)
+
+**Note:** Version bump only for package @ttoss/google-maps
+
+## [1.22.9](https://github.com/ttoss/ttoss/compare/@ttoss/google-maps@1.22.8...@ttoss/google-maps@1.22.9) (2023-03-22)
+
+**Note:** Version bump only for package @ttoss/google-maps
+
+## [1.22.8](https://github.com/ttoss/ttoss/compare/@ttoss/google-maps@1.22.7...@ttoss/google-maps@1.22.8) (2023-03-20)
+
+**Note:** Version bump only for package @ttoss/google-maps
+
+## [1.22.7](https://github.com/ttoss/ttoss/compare/@ttoss/google-maps@1.22.6...@ttoss/google-maps@1.22.7) (2023-03-16)
+
+**Note:** Version bump only for package @ttoss/google-maps
+
 ## [1.22.6](https://github.com/ttoss/ttoss/compare/@ttoss/google-maps@1.22.5...@ttoss/google-maps@1.22.6) (2023-03-07)
 
 **Note:** Version bump only for package @ttoss/google-maps