+++ packages/landing-pages/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.16](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.15...@ttoss/landing-pages@0.6.16) (2023-03-26)
+
+**Note:** Version bump only for package @ttoss/landing-pages
+
+## [0.6.15](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.14...@ttoss/landing-pages@0.6.15) (2023-03-26)
+
+**Note:** Version bump only for package @ttoss/landing-pages
+
+## [0.6.14](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.13...@ttoss/landing-pages@0.6.14) (2023-03-22)
+
+### Bug Fixes
+
+- add type button as default ([#252](https://github.com/ttoss/ttoss/issues/252)) ([6429e7c](https://github.com/ttoss/ttoss/commit/6429e7c30e286a98316de9b109129c1b094f6680))
+
+## [0.6.13](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.12...@ttoss/landing-pages@0.6.13) (2023-03-20)
+
+**Note:** Version bump only for package @ttoss/landing-pages
+
+## [0.6.12](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.11...@ttoss/landing-pages@0.6.12) (2023-03-17)
+
+**Note:** Version bump only for package @ttoss/landing-pages
+
+## [0.6.11](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.10...@ttoss/landing-pages@0.6.11) (2023-03-16)
+
+**Note:** Version bump only for package @ttoss/landing-pages
+
 ## [0.6.10](https://github.com/ttoss/ttoss/compare/@ttoss/landing-pages@0.6.9...@ttoss/landing-pages@0.6.10) (2023-03-15)
 
 **Note:** Version bump only for package @ttoss/landing-pages