+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.3](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.2...@ttoss/theme@1.4.3) (2023-03-26)
+
+### Bug Fixes
+
+- @ttoss/theme exports ([3fd2243](https://github.com/ttoss/ttoss/commit/3fd2243d8dbb2735ba8126d5da7604a0629ef9cd))
+
+## [1.4.2](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.1...@ttoss/theme@1.4.2) (2023-03-26)
+
+**Note:** Version bump only for package @ttoss/theme
+
+## [1.4.1](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.4.0...@ttoss/theme@1.4.1) (2023-03-22)
+
+**Note:** Version bump only for package @ttoss/theme
+
+# [1.4.0](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.3.0...@ttoss/theme@1.4.0) (2023-03-20)
+
+### Features
+
+- add bruttal theme icons ([#249](https://github.com/ttoss/ttoss/issues/249)) ([df3d245](https://github.com/ttoss/ttoss/commit/df3d245901ea806b847bcce4af627c2c94fa12b3))
+
+# [1.3.0](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.2.10...@ttoss/theme@1.3.0) (2023-03-17)
+
+### Features
+
+- add root styles ([#247](https://github.com/ttoss/ttoss/issues/247)) ([d6381e5](https://github.com/ttoss/ttoss/commit/d6381e51f74cabae4b5b8b49568fa1c3622a955b))
+
+## [1.2.10](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.2.9...@ttoss/theme@1.2.10) (2023-03-16)
+
+**Note:** Version bump only for package @ttoss/theme
+
 ## [1.2.9](https://github.com/ttoss/ttoss/compare/@ttoss/theme@1.2.8...@ttoss/theme@1.2.9) (2023-03-15)
 
 **Note:** Version bump only for package @ttoss/theme