+++ terezinha-farm/auth/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.21](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.20...@terezinha-farm/auth@0.2.21) (2023-03-26)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.20](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.19...@terezinha-farm/auth@0.2.20) (2023-03-22)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.19](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.18...@terezinha-farm/auth@0.2.19) (2023-03-21)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.18](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.17...@terezinha-farm/auth@0.2.18) (2023-03-20)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.17](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.16...@terezinha-farm/auth@0.2.17) (2023-03-19)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
+## [0.2.16](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.15...@terezinha-farm/auth@0.2.16) (2023-03-16)
+
+**Note:** Version bump only for package @terezinha-farm/auth
+
 ## [0.2.15](https://github.com/ttoss/ttoss/compare/@terezinha-farm/auth@0.2.14...@terezinha-farm/auth@0.2.15) (2023-03-07)
 
 **Note:** Version bump only for package @terezinha-farm/auth