+++ terezinha-farm/iam/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.17](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.16...@terezinha-farm/iam@0.2.17) (2023-03-26)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.16](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.15...@terezinha-farm/iam@0.2.16) (2023-03-22)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.15](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.14...@terezinha-farm/iam@0.2.15) (2023-03-21)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.14](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.13...@terezinha-farm/iam@0.2.14) (2023-03-20)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.13](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.12...@terezinha-farm/iam@0.2.13) (2023-03-19)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
+## [0.2.12](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.11...@terezinha-farm/iam@0.2.12) (2023-03-16)
+
+**Note:** Version bump only for package @terezinha-farm/iam
+
 ## [0.2.11](https://github.com/ttoss/ttoss/compare/@terezinha-farm/iam@0.2.10...@terezinha-farm/iam@0.2.11) (2023-03-07)
 
 **Note:** Version bump only for package @terezinha-farm/iam